{{cookiecutter.project_slug}}/requirements/base.txt
{% if cookiecutter.use_python3 == 'y' -%}
# Wheel 0.25+ needed to install certain packages on CPython 3.5+
# like Pillow and psycopg2
# See http://bitly.com/wheel-building-fails-CPython-35
# Verified bug on Python 3.5.1
wheel==0.29.0
{%- endif %}

# Bleeding edge Django
django==1.9.4

# Configuration
django-environ==0.4.0

# Forms
django-braces==1.10.0
django-crispy-forms==1.6.1

# Models
django-model-utils==2.6

# Images
Pillow==3.4.2

# Python-PostgreSQL Database Adapter
psycopg2==2.6.2

# Unicode slugification
awesome-slugify==1.6.5

# Time zones support
pytz==2016.7

# Redis support
django-redis==4.5.0
redis>=2.10.5

{% if cookiecutter.use_celery == "y" %}
celery==3.1.24
{% endif %}

# Your custom requirements go here
-e git://github.com/sshwsfc/django-xadmin.git@#egg=django-xadmin
django_formtools==1.0
-e https://github.com/toastdriven/restless/
django-cacheops==2.4.3
requests==2.11.1
